updated life damage attack for enemy
added new enemy to spwnpoints nad other stuff

+++ New Unity Game/Assets/scripts/strongholdScript.cs
@@ -3,7 +3,7 @@ using System.Collections;
 
 public class strongholdScript : MonoBehaviour {
 
-	public int defence;
+	public float defence;
 	public GameObject player;
 	public GameObject exp;
